Refactord Ballot Leader Election Added get_leader() to retrieve the current leader in BLE Fixed some of the comments Moved BLEMessages and BLE to leader_election module
