--- src/ble/mod.rs
@@ -1,3 +1,0 @@
-/// Ballot Leader Election algorithm for electing new leaders
-pub mod ballot_leader_election;
-mod messages;
